CJA Get Audit Logs

This function will pull a list of audit logs defined by the different defined parameters.

cja_get_audit_logs(
  startDate = NULL,
  endDate = NULL,
  action = NULL,
  component = NULL,
  componentId = NULL,
  userType = NULL,
  userId = NULL,
  userEmail = NULL,
  description = NULL,
  pageSize = 100,
  pageNumber = 0,
  debug = FALSE
)

Arguments

startDate

Date is not required, but if you filter by date, both start & end date must be set.

endDate

Date is not required, but if you filter by date, both start & end date must be set.

action

The action you want to filter by.See details section for options

component

The type of component you want to filter by. See details section for options

componentId

The ID of the component.

userType

The type of user.

userId

The ID of the user.

userEmail

The email address of the user.

description

The log description you want to filter by.

pageSize

Number of results per page. If left null, the default size will be set to 100.

pageNumber

Page number (base 0 - first page is "0")

debug

Used to help troubleshoot api call issues. Shows the call and result in the console

Value

A data frame of audit logs and corresponding metadata

Details

startDate/endDate format

Action available values are: 'CREATE', 'EDIT', 'DELETE', 'LOGIN_FAILED', 'LOGIN_SUCCESSFUL', 'API_REQUEST', 'LOGOUT', 'APPROVE', 'UNAPPROVE', 'SHARE', 'UNSHARE', 'TRANSFER', 'ORG_CHANGE'

Component available values are: 'ANNOTATION', 'CALCULATED_METRIC', 'CONNECTION', 'DATA_GROUP', 'DATA_VIEW', 'DATE_RANGE', 'FILTER', 'MOBILE', 'PROJECT', 'REPORT', 'SCHEDULED_PROJECT', 'USER', 'USER_GROUP', 'IMS_ORG', 'FEATURE_ACCESS'

Examples

if (FALSE) {
cja_get_audit_logs()
}