Get a paginated list of filters in CJA

Retrieves a paginated list of filters, also known as segments in Adobe Analytics.

cja_get_filters(
  expansion = NULL,
  includeType = "all",
  dataviewIds = NULL,
  ownerId = NULL,
  filterByIds = NULL,
  toBeUsedInRsid = NULL,
  locale = "en_US",
  name = NULL,
  filterByModifiedAfter = NULL,
  cached = TRUE,
  pagination = TRUE,
  limit = 10,
  page = 0,
  sortDirection = "ASC",
  sortProperty = "id",
  debug = FALSE
)

Arguments

expansion

Comma-delimited list of additional segment metadata fields to include on response. See Details for all options available

includeType

Include additional filters not owned by user. Default is "all". Options include: "shared" "templates" "deleted" "internal"

dataviewIds

Filter list to only include filters tied to the specified data group ID list (comma-delimited)

ownerId

Filter list to only include filters owned by the specified imsUserId

filterByIds

Filter list to only include filters in the specified list (comma-delimited list of IDs). This has filtered Ids from tags, approved, favorites and user specified Ids list.

toBeUsedInRsid

The report suite where the segment is intended to be used. This report suite will be used to determine things like compatibility and permissions.

locale

Locale - Default: "en_US"

name

Filter list to only include filters that contains the Name. Can only be a string value.

filterByModifiedAfter

Filter list to only include filters modified since this date. 'yyyy-mm-dd' format

cached

Return cached results. TRUE by default.

pagination

Return paginated results

limit

Number of results per page

page

Page number (base 0 - first page is "0")

sortDirection

Sort direction ('ASC' or 'DESC'). 'ASC' is default.

sortProperty

Property to sort by (name, modified_date, performanceScore, id is currently allowed). 'id' is default

debug

Used to help troubleshoot api call issues. Shows the call and result in the console

Value

A data frame of company ids and company names

Details

Expansion options can include the following: "compatibility", "definition", "internal", "modified", "isDeleted", "definitionLastModified", "createdDate", "recentRecordedAccess", "performanceScore", "owner", "dataId", "ownerFullName", "dataName", "sharesFullName", "approved", "favorite", "shares", "tags", "usageSummary", "usageSummaryWithRelevancyScore"

Examples

if (FALSE) {
cja_get_filters()
}